// src/app/models/usuarios.model.ts
export class Usuario {
  id: string | undefined;
  nombres: string | undefined;
  apellidos: string | undefined;
  fechaNacimiento: string | undefined;
  email: string | undefined;
  numDoc: number | undefined;
  area: number | undefined;
  salario: number | undefined;
  estado: boolean | undefined;
  password?: string;
}